+++ pkg/ccr/job.go
@@ -341,7 +341,7 @@ func (j *Job) fullSync() error {
 
 		if j.SyncType == TableSync {
 			if _, ok := tableCommitSeqMap[j.Src.TableId]; !ok {
-				return xerror.Errorf(xerror.Normal, "tableid %d, commit seq not found", j.Src.TableId)
+				return xerror.Errorf(xerror.Normal, "table id %d, commit seq not found", j.Src.TableId)
 			}
 		}
 
@@ -361,6 +361,9 @@ func (j *Job) fullSync() error {
 		jobInfo := snapshotResp.GetJobInfo()
 		tableCommitSeqMap := inMemoryData.TableCommitSeqMap
 
+		log.Infof("snapshot response meta size: %d, job info size: %d",
+			len(snapshotResp.Meta), len(snapshotResp.JobInfo))
+
 		var jobInfoMap map[string]interface{}
 		err := json.Unmarshal(jobInfo, &jobInfoMap)
 		if err != nil {
@@ -379,7 +382,7 @@ func (j *Job) fullSync() error {
 		if err != nil {
 			return xerror.Errorf(xerror.Normal, "marshal jobInfo failed, jobInfo: %v", jobInfoMap)
 		}
-		log.Debugf("jobInfoBytes: %s", string(jobInfoBytes))
+		log.Debugf("job info size: %d, bytes: %s", len(jobInfoBytes), string(jobInfoBytes))
 		snapshotResp.SetJobInfo(jobInfoBytes)
 
 		var commitSeq int64 = math.MaxInt64
@@ -455,9 +458,7 @@ func (j *Job) fullSync() error {
 				}
 				log.Infof("the signature of table %s is not matched with the target table in snapshot", tableName)
 				for {
-					dropSql := fmt.Sprintf("DROP TABLE %s FORCE", utils.FormatKeywordName(tableName))
-					log.Infof("drop table sql: %s", dropSql)
-					if err := j.destMeta.DbExec(dropSql); err == nil {
+					if err := j.IDest.DropTable(tableName); err == nil {
 						break
 					}
 				}
@@ -486,6 +487,13 @@ func (j *Job) fullSync() error {
 				srcTableName, err := j.srcMeta.GetTableNameById(srcTableId)
 				if err != nil {
 					return err
+				}
+
+				// If srcTableName is empty, it may be deleted.
+				// No need to map it to dest table
+				if srcTableName == "" {
+					log.Warnf("the name of source table id: %d is empty, no need to map it to dest table", srcTableId)
+					continue
 				}
 
 				destTableId, err := j.destMeta.GetTableId(srcTableName)
@@ -720,10 +728,10 @@ func (j *Job) handleUpsert(binlog *festruct.TBinlog) error {
 		// Step 1: get related tableRecords
 		tableRecords, err := j.getReleatedTableRecords(upsert)
 		if err != nil {
-			log.Errorf("get releated table records failed, err: %+v", err)
+			log.Errorf("get related table records failed, err: %+v", err)
 		}
 		if len(tableRecords) == 0 {
-			log.Debug("no releated table records")
+			log.Debug("no related table records")
 			return nil
 		}
 
@@ -788,6 +796,7 @@ func (j *Job) handleUpsert(binlog *festruct.TBinlog) error {
 		commitInfos, err := j.ingestBinlog(txnId, tableRecords)
 		if err != nil {
 			rollback(err, inMemoryData)
+			return err
 		} else {
 			log.Debugf("commitInfos: %v", commitInfos)
 			inMemoryData.CommitInfos = commitInfos
@@ -882,6 +891,11 @@ func (j *Job) handleAddPartition(binlog *festruct.TBinlog) error {
 		return err
 	}
 
+	if addPartition.IsTemp {
+		log.Infof("skip add temporary partition because backup/restore table with temporary partitions is not supported yet")
+		return nil
+	}
+
 	var destTableName string
 	if j.SyncType == TableSync {
 		destTableName = j.Dest.Table
@@ -897,10 +911,7 @@ func (j *Job) handleAddPartition(binlog *festruct.TBinlog) error {
 			return xerror.Errorf(xerror.Normal, "tableId %d not found in destMeta", destTableId)
 		}
 	}
-
-	addPartitionSql := addPartition.GetSql(destTableName)
-	log.Infof("addPartitionSql: %s", addPartitionSql)
-	return j.IDest.DbExec(addPartitionSql)
+	return j.IDest.AddPartition(destTableName, addPartition)
 }
 
 // handleDropPartition
@@ -913,7 +924,6 @@ func (j *Job) handleDropPartition(binlog *festruct.TBinlog) error {
 		return err
 	}
 
-	destDbName := j.Dest.Database
 	var destTableName string
 	if j.SyncType == TableSync {
 		destTableName = j.Dest.Table
@@ -929,11 +939,7 @@ func (j *Job) handleDropPartition(binlog *festruct.TBinlog) error {
 			return xerror.Errorf(xerror.Normal, "tableId %d not found in destMeta", destTableId)
 		}
 	}
-
-	// dropPartitionSql = "ALTER TABLE " + sql
-	dropPartitionSql := fmt.Sprintf("ALTER TABLE %s.%s %s", utils.FormatKeywordName(destDbName), utils.FormatKeywordName(destTableName), dropPartition.Sql)
-	log.Infof("dropPartitionSql: %s", dropPartitionSql)
-	return j.IDest.Exec(dropPartitionSql)
+	return j.IDest.DropPartition(destTableName, dropPartition)
 }
 
 // handleCreateTable
@@ -950,11 +956,8 @@ func (j *Job) handleCreateTable(binlog *festruct.TBinlog) error {
 		return err
 	}
 
-	sql := createTable.Sql
-	log.Infof("createTableSql: %s", sql)
-	// HACK: for drop table
-	if err := j.IDest.DbExec(sql); err != nil {
-		return err
+	if err := j.IDest.CreateTableOrView(createTable, j.Src.Database); err != nil {
+		return xerror.Wrapf(err, xerror.Normal, "create table %d", createTable.TableId)
 	}
 
 	j.srcMeta.GetTables()
@@ -993,7 +996,7 @@ func (j *Job) handleDropTable(binlog *festruct.TBinlog) error {
 	}
 
 	tableName := dropTable.TableName
-	// depreated
+	// deprecated
 	if tableName == "" {
 		dirtySrcTables := j.srcMeta.DirtyGetTables()
 		srcTable, ok := dirtySrcTables[dropTable.TableId]
@@ -1004,10 +1007,8 @@ func (j *Job) handleDropTable(binlog *festruct.TBinlog) error {
 		tableName = srcTable.Name
 	}
 
-	sql := fmt.Sprintf("DROP TABLE %s FORCE", utils.FormatKeywordName(tableName))
-	log.Infof("dropTableSql: %s", sql)
-	if err = j.IDest.DbExec(sql); err != nil {
-		return err
+	if err = j.IDest.DropTable(tableName); err != nil {
+		return xerror.Wrapf(err, xerror.Normal, "drop table %s", tableName)
 	}
 
 	j.srcMeta.GetTables()
@@ -1043,17 +1044,39 @@ func (j *Job) handleAlterJob(binlog *festruct.TBinlog) error {
 		return nil
 	}
 
-	for {
-		// drop table dropTableSql
-		var dropTableSql string
-		if j.SyncType == TableSync {
-			dropTableSql = fmt.Sprintf("DROP TABLE %s FORCE", utils.FormatKeywordName(j.Dest.Table))
-		} else {
-			dropTableSql = fmt.Sprintf("DROP TABLE %s FORCE", utils.FormatKeywordName(alterJob.TableName))
-		}
-		log.Infof("dropTableSql: %s", dropTableSql)
+	// drop table dropTableSql
+	var destTableName string
+	if j.SyncType == TableSync {
+		destTableName = j.Dest.Table
+	} else {
+		destTableName = alterJob.TableName
+	}
 
-		if err := j.destMeta.DbExec(dropTableSql); err == nil {
+	var allViewDeleted bool = false
+	for {
+		// before drop table, drop related view firstly
+		if !allViewDeleted {
+			views, err := j.IDest.GetAllViewsFromTable(destTableName)
+			if err != nil {
+				log.Errorf("when alter job, get view from table failed, err : %v", err)
+				continue
+			}
+
+			var dropViewFailed bool = false
+			for _, view := range views {
+				if err := j.IDest.DropView(view); err != nil {
+					log.Errorf("when alter job, drop view %s failed, err : %v", view, err)
+					dropViewFailed = true
+				}
+			}
+			if dropViewFailed {
+				continue
+			}
+
+			allViewDeleted = true
+		}
+
+		if err := j.IDest.DropTable(destTableName); err == nil {
 			break
 		}
 	}
@@ -1071,19 +1094,7 @@ func (j *Job) handleLightningSchemaChange(binlog *festruct.TBinlog) error {
 		return err
 	}
 
-	log.Debugf("lightningSchemaChange %v", lightningSchemaChange)
-
-	rawSql := lightningSchemaChange.RawSql
-	//   "rawSql": "ALTER TABLE `default_cluster:ccr`.`test_ddl` ADD COLUMN `nid1` int(11) NULL COMMENT \"\""
-	// replace `default_cluster:${Src.Database}`.`test_ddl` to `test_ddl`
-	var sql string
-	if strings.Contains(rawSql, fmt.Sprintf("`default_cluster:%s`.", j.Src.Database)) {
-		sql = strings.Replace(rawSql, fmt.Sprintf("`default_cluster:%s`.", j.Src.Database), "", 1)
-	} else {
-		sql = strings.Replace(rawSql, fmt.Sprintf("`%s`.", j.Src.Database), "", 1)
-	}
-	log.Infof("lightningSchemaChangeSql, rawSql: %s, sql: %s", rawSql, sql)
-	return j.IDest.DbExec(sql)
+	return j.IDest.LightningSchemaChange(j.Src.Database, lightningSchemaChange)
 }
 
 func (j *Job) handleTruncateTable(binlog *festruct.TBinlog) error {
@@ -1105,16 +1116,7 @@ func (j *Job) handleTruncateTable(binlog *festruct.TBinlog) error {
 		return xerror.Panicf(xerror.Normal, "invalid sync type: %v", j.SyncType)
 	}
 
-	var sql string
-	if truncateTable.RawSql == "" {
-		sql = fmt.Sprintf("TRUNCATE TABLE %s", utils.FormatKeywordName(destTableName))
-	} else {
-		sql = fmt.Sprintf("TRUNCATE TABLE %s %s", utils.FormatKeywordName(destTableName), truncateTable.RawSql)
-	}
-
-	log.Infof("truncateTableSql: %s", sql)
-
-	err = j.IDest.DbExec(sql)
+	err = j.IDest.TruncateTable(destTableName, truncateTable)
 	if err == nil {
 		if srcTableName, err := j.srcMeta.GetTableNameById(truncateTable.TableId); err == nil {
 			// if err != nil, maybe truncate table had been dropped
@@ -1124,6 +1126,14 @@ func (j *Job) handleTruncateTable(binlog *festruct.TBinlog) error {
 	}
 
 	return err
+}
+
+func (j *Job) handleReplacePartitions(binlog *festruct.TBinlog) error {
+	log.Infof("handle replace partitions binlog, commit seq: %d", *binlog.CommitSeq)
+
+	// TODO(walter) replace partitions once backuping/restoring with temporary partitions is supportted.
+
+	return j.newSnapshot(j.progress.CommitSeq)
 }
 
 // return: error && bool backToRunLoop
@@ -1203,6 +1213,8 @@ func (j *Job) handleBinlog(binlog *festruct.TBinlog) error {
 		log.Info("handle barrier binlog, ignore it")
 	case festruct.TBinlogType_TRUNCATE_TABLE:
 		return j.handleTruncateTable(binlog)
+	case festruct.TBinlogType_REPLACE_PARTITIONS:
+		return j.handleReplacePartitions(binlog)
 	default:
 		return xerror.Errorf(xerror.Normal, "unknown binlog type: %v", binlog.GetType())
 	}
@@ -1223,7 +1235,8 @@ func (j *Job) recoverIncrementalSync() error {
 
 func (j *Job) incrementalSync() error {
 	if !j.progress.IsDone() {
-		log.Infof("job progress is not done, state is (%s), need recover", j.progress.SubSyncState)
+		log.Infof("job progress is not done, need recover. state: %s, prevCommitSeq: %d, commitSeq: %d",
+			j.progress.SubSyncState, j.progress.PrevCommitSeq, j.progress.CommitSeq)
 
 		return j.recoverIncrementalSync()
 	}
@@ -1239,6 +1252,7 @@ func (j *Job) incrementalSync() error {
 
 	// Step 2: handle all binlog
 	for {
+		// The CommitSeq is equals to PrevCommitSeq in here.
 		commitSeq := j.progress.CommitSeq
 		log.Debugf("src: %s, commitSeq: %v", src, commitSeq)
 
@@ -1365,6 +1379,7 @@ func (j *Job) handleError(err error) error {
 	}
 
 	if xerr.Category() == xerror.Meta {
+		log.Warnf("receive meta category error, make new snapshot, job: %s, err: %v", j.Name, err)
 		j.newSnapshot(j.progress.CommitSeq)
 	}
 	return nil
@@ -1476,37 +1491,23 @@ func (j *Job) desyncTable() error {
 	if err != nil {
 		return err
 	}
-
-	desyncSql := fmt.Sprintf("ALTER TABLE %s SET (\"is_being_synced\"=\"false\")", tableName)
-	log.Debugf("db exec: %s", desyncSql)
-	if err := j.IDest.DbExec(desyncSql); err != nil {
-		return xerror.Wrapf(err, xerror.FE, "failed tables: %s", tableName)
-	}
-	return nil
+	return j.IDest.DesyncTables(tableName)
 }
 
 func (j *Job) desyncDB() error {
 	log.Debugf("desync db")
 
-	var failedTable string = ""
 	tables, err := j.destMeta.GetTables()
 	if err != nil {
 		return err
 	}
 
+	tableNames := []string{}
 	for _, tableMeta := range tables {
-		desyncSql := fmt.Sprintf("ALTER TABLE %s SET (\"is_being_synced\"=\"false\")", tableMeta.Name)
-		log.Debugf("db exec: %s", desyncSql)
-		if err := j.IDest.DbExec(desyncSql); err != nil {
-			failedTable += tableMeta.Name + " "
-		}
+		tableNames = append(tableNames, tableMeta.Name)
 	}
 
-	if failedTable != "" {
-		return xerror.Errorf(xerror.FE, "failed tables: %s", failedTable)
-	}
-
-	return nil
+	return j.IDest.DesyncTables(tableNames...)
 }
 
 func (j *Job) Desync() error {