+++ pkg/ccr/job_progress.go
@@ -136,6 +136,7 @@ type JobProgress struct {
 	// Sub sync state machine states
 	SubSyncState SubSyncState `json:"sub_sync_state"`
 
+	// The commit seq where the target cluster has synced.
 	PrevCommitSeq     int64           `json:"prev_commit_seq"`
 	CommitSeq         int64           `json:"commit_seq"`
 	TableMapping      map[int64]int64 `json:"table_mapping"`
@@ -210,7 +211,6 @@ func NewJobProgressFromJson(jobName string, db storage.DB) (*JobProgress, error)
 }
 
 func (j *JobProgress) StartHandle(commitSeq int64) {
-	j.PrevCommitSeq = j.CommitSeq
 	j.CommitSeq = commitSeq
 
 	j.Persist()
@@ -265,17 +265,24 @@ func (j *JobProgress) CommitNextSubWithPersist(commitSeq int64, subSyncState Sub
 	j.Persist()
 }
 
+// Switch to new sync state.
+//
+// The PrevCommitSeq is set to commitSeq, if the sub sync state is done.
 func (j *JobProgress) NextWithPersist(commitSeq int64, syncState SyncState, subSyncState SubSyncState, persistData string) {
 	if subSyncState == BeginCreateSnapshot && (syncState == TableFullSync || syncState == DBFullSync) {
 		j.FullSyncStartAt = time.Now().Unix()
 		j.IncrementalSyncStartAt = 0
 		j.IngestBinlogAt = 0
-	} else if subSyncState == Done && (syncState == TableIncrementalSync || syncState == TableIncrementalSync) {
+	} else if subSyncState == Done && (syncState == TableIncrementalSync || syncState == DBIncrementalSync) {
 		j.IncrementalSyncStartAt = time.Now().Unix()
 		j.IngestBinlogAt = 0
 	}
 
 	j.CommitSeq = commitSeq
+	if subSyncState == Done {
+		j.PrevCommitSeq = commitSeq
+	}
+
 	j.SyncState = syncState
 	j.SubSyncState = subSyncState
 	j.PersistData = persistData
@@ -284,7 +291,7 @@ func (j *JobProgress) NextWithPersist(commitSeq int64, syncState SyncState, subS
 	j.Persist()
 }
 
-func (j *JobProgress) IsDone() bool { return j.SubSyncState == Done }
+func (j *JobProgress) IsDone() bool { return j.SubSyncState == Done && j.PrevCommitSeq == j.CommitSeq }
 
 // TODO(Drogon): check reset some fields
 func (j *JobProgress) Done() {
@@ -338,5 +345,6 @@ func (j *JobProgress) Persist() {
 		break
 	}
 
-	log.Trace("update job progress done")
+	log.Tracef("update job progress done, state: %s, subState: %s, commitSeq: %d, prevCommitSeq: %d",
+		j.SyncState, j.SubSyncState, j.CommitSeq, j.PrevCommitSeq)
 }